3rd commit
Looks good. Just need to make sure tweets are in order (a minor thing but want to be consistent). May want to consider making a list of tuples with each tuple being the created_at and the tweet. Would help with sanity checks as well as. Nice plot too.

Need to clean up tags